Refactor common HTTP proxy code out to OkHttpClientHelper
